requirements.txt
maskpass==0.3.7
pycryptodome==3.17
PyInstaller==5.8.0
requests==2.28.2
